fixes #49 an auto transition can now use transition linkage parameter to transition to the target resource.
Move the link path mapping from transition create time to link creation time.  Prior to this change we would create a transition with a modified path based on the linkage parameters.  This way we preserve the target path and delay the link creation until we need to create the link.